pyproject.toml
[tool.poetry]
name = "megai-man"
version = "0.1.0"
description = ""
authors = ["Victor Severo <victor.severo.miranda@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
stable-retro = "^0.9.2"
stable-baselines3 = "^2.3.2"
opencv-python = "^4.9.0.80"
tensorboard = "^2.15.1"
tensorboard-plugin-profile = "^2.15.0"
torch-tb-profiler = "^0.4.3"
pygame = "^2.5.2"
sb3-contrib = "^2.3.0"
wandb = "^0.17.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
flake8 = "^7.0.0"
isort = "^5.13.2"
notebook = "^7.0.6"
ipywidgets = "^8.1.1"
seaborn = "^0.12.2"
optuna = "^3.5.0"
plotly = "^5.18.0"
optuna-dashboard = "^0.14.0"
torchvision = "^0.17.1"
snakeviz = "^2.2.0"
rich = "^13.7.1"
torchviz = "^0.0.2"
torch-summary = "^1.4.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
